Feature: A LazyValueMap behaving like a regular map with lazy values. (#275)
A Map that can store its values in a Lazy manner.

This has the advantage that while providing 'standard' Map features,
unused values do not need to be evaluated.

The behaviour of this map depends on the delegate map it was initialized with.

If the delegate map is mutable, the lazy map will be mutable as well.
If the delegate map is sorted, the lazy map will be sorted as well.
The default and copy constructors create a lazy map that behaves like a LinkedHashMap.
If the backing map does not support null values, the lazy map will support them.
This fixes #274

---------

Signed-off-by: Sjoerd Talsma <sjoerdtalsma@users.noreply.github.com>